Edit text editing module API & make tests pass
also, use minified wysihtml5 js file instead of unminified one. Not
doing that breaks browserify because the comments in the wysihtml5
source have <script> tags in multiline comments, which for some reason
breaks inline scripts. god.

on that note, we should switch to https://github.com/guardian/scribe
soon because wysihtml5 is not being developed anymore and it's not
ideal.