Tweak implementation of common model behaviors
Previously, there was a lot of repetition in model and view definitions
for working with related topics. That was not great and would have
surely caused problems. Additionally, the code for requiring a model to
have a defined project had to be inherited and prevented us from
overriding the constructor.

This might not be perfect, but it works for now. The two "mixins" for
related topics both have a "constructor" property which must be called
explicitly and then a "mixin" property which is extended with the
`backbone.cocktail` package.