Render components on server with react
Routes are required to provide a Component attribute which instructs
which component should be rendered in the "content" part of the
application. They can also provide a getData method which can fetch
data from the API (and only from the API). Any other data fetching
(from other servers, localStorage, indexedDB, etc.) should be done
solely by the client in componentDidMount.